app: move POST /api/release to app.routers.release
